rename private/stdint.h -> private/mintomic_stdint.h
